refactor: align academic repository identifiers to long

+++ EduOS.Core/Interfaces/IRepositories/IClassRepository.cs
@@ -4,9 +4,9 @@ namespace EduOS.Core.Interfaces.IRepositories
 {
     public interface IClassRepository : IGenericRepository<Class>
     {
-        Task<bool> IsClassNameExistsAsync(string name, int tenantId, int? excludeId = null);
-        Task<List<Class>> GetActiveClassesAsync(int tenantId);
-        Task<Class?> GetWithSectionsAsync(int id);
-        Task<Class?> GetWithSubjectsAsync(int id);
+        Task<bool> IsClassNameExistsAsync(string name, long tenantId, long? excludeId = null);
+        Task<List<Class>> GetActiveClassesAsync(long tenantId);
+        Task<Class?> GetWithSectionsAsync(long id);
+        Task<Class?> GetWithSubjectsAsync(long id);
     }
 }
